add template for magic code email
